refac: changed references to new folder
