+++ source/geh_settlement_report/tests/settlement_reports/application/tasks/test_execute_wholesale_results.py
@@ -15,7 +15,7 @@ from geh_settlement_report.settlement_reports.domain.utils.market_role import Ma
 from geh_settlement_report.settlement_reports.domain.utils.report_data_type import ReportDataType
 from geh_settlement_report.settlement_reports.infrastructure.paths import get_report_output_path
 from tests.assertion import assert_file_names_and_columns
-from tests.data_seeding.standard_wholesale_fixing_scenario_data_generator import (
+from tests.settlement_reports.data_seeding.standard_wholesale_fixing_scenario_data_generator import (
     CHARGE_OWNER_ID_WITHOUT_TAX,
 )
 from tests.utils import (
