+++ source/geh_settlement_report/tests/settlement_reports/domain/charge_link_periods/test_charge_link_periods_read_and_filter.py
@@ -6,10 +6,10 @@ from unittest.mock import Mock
 import pytest
 from pyspark.sql import DataFrame, SparkSession
 
-import tests.test_factories.charge_link_periods_factory as charge_link_periods_factory
-import tests.test_factories.charge_price_information_periods_factory as charge_price_information_periods_factory
-import tests.test_factories.default_test_data_spec as default_data
-import tests.test_factories.metering_point_periods_factory as metering_point_periods_factory
+import tests.settlement_reports.test_factories.charge_link_periods_factory as charge_link_periods_factory
+import tests.settlement_reports.test_factories.charge_price_information_periods_factory as charge_price_information_periods_factory
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.metering_point_periods_factory as metering_point_periods_factory
 from geh_settlement_report.settlement_reports.domain.charge_link_periods.read_and_filter import (
     read_and_filter,
 )
