+++ source/geh_settlement_report/tests/settlement_reports/domain/metering_point_periods/test_metering_point_periods_factory_balance_fixing.py
@@ -3,8 +3,8 @@ from unittest.mock import Mock
 
 from pyspark.sql import DataFrame, SparkSession
 
-import tests.test_factories.default_test_data_spec as default_data
-import tests.test_factories.metering_point_periods_factory as input_metering_point_periods_factory
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.metering_point_periods_factory as input_metering_point_periods_factory
 from geh_settlement_report.settlement_reports.application.job_args.settlement_report_args import (
     SettlementReportArgs,
 )
@@ -12,7 +12,7 @@ from geh_settlement_report.settlement_reports.domain.metering_point_periods.mete
     create_metering_point_periods,
 )
 from geh_settlement_report.settlement_reports.domain.utils.csv_column_names import CsvColumnNames
-from tests.test_factories import latest_calculations_factory
+from tests.settlement_reports.test_factories import latest_calculations_factory
 from tests.utils import Dates as d
