+++ source/geh_settlement_report/tests/settlement_reports/domain/metering_point_periods/test_metering_point_periods_factory_wholesale.py
@@ -3,10 +3,10 @@ from unittest.mock import Mock
 
 from pyspark.sql import DataFrame, SparkSession
 
-import tests.test_factories.charge_link_periods_factory as input_charge_link_periods_factory
-import tests.test_factories.charge_price_information_periods_factory as input_charge_price_information_periods_factory
-import tests.test_factories.default_test_data_spec as default_data
-import tests.test_factories.metering_point_periods_factory as input_metering_point_periods_factory
+import tests.settlement_reports.test_factories.charge_link_periods_factory as input_charge_link_periods_factory
+import tests.settlement_reports.test_factories.charge_price_information_periods_factory as input_charge_price_information_periods_factory
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.metering_point_periods_factory as input_metering_point_periods_factory
 from geh_settlement_report.settlement_reports.application.job_args.settlement_report_args import (
     SettlementReportArgs,
 )
