+++ source/geh_settlement_report/tests/settlement_reports/domain/metering_point_periods/test_metering_point_periods_read_and_filter_balance_fixing.py
@@ -3,14 +3,14 @@ from unittest.mock import Mock
 
 from pyspark.sql import DataFrame, SparkSession
 
-import tests.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
 from geh_settlement_report.settlement_reports.domain.metering_point_periods.read_and_filter_balance_fixing import (
     read_and_filter,
 )
 from geh_settlement_report.settlement_reports.infrastructure.wholesale.column_names import (
     DataProductColumnNames,
 )
-from tests.test_factories import (
+from tests.settlement_reports.test_factories import (
     latest_calculations_factory,
     metering_point_periods_factory,
 )
