+++ source/geh_settlement_report/tests/settlement_reports/domain/monthly_amounts/test_monthly_amounts_prepare_for_csv.py
@@ -2,8 +2,8 @@ import pytest
 from pyspark.sql import SparkSession
 from pyspark.sql.functions import lit
 
-import tests.test_factories.default_test_data_spec as default_data
-import tests.test_factories.monthly_amounts_per_charge_factory as monthly_amounts_per_charge_factory
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.monthly_amounts_per_charge_factory as monthly_amounts_per_charge_factory
 from geh_settlement_report.settlement_reports.domain.monthly_amounts.prepare_for_csv import (
     prepare_for_csv,
 )
