+++ source/geh_settlement_report/tests/settlement_reports/domain/monthly_amounts/test_monthly_amounts_read_and_filter.py
@@ -4,9 +4,9 @@ import pytest
 from pyspark.sql import SparkSession
 from pyspark.sql import functions as F
 
-import tests.test_factories.default_test_data_spec as default_data
-import tests.test_factories.monthly_amounts_per_charge_factory as monthly_amounts_per_charge_factory
-import tests.test_factories.total_monthly_amounts_factory as total_monthly_amounts_factory
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.monthly_amounts_per_charge_factory as monthly_amounts_per_charge_factory
+import tests.settlement_reports.test_factories.total_monthly_amounts_factory as total_monthly_amounts_factory
 from geh_settlement_report.settlement_reports.application.job_args.settlement_report_args import (
     SettlementReportArgs,
 )
@@ -18,7 +18,7 @@ from geh_settlement_report.settlement_reports.domain.utils.market_role import Ma
 from geh_settlement_report.settlement_reports.infrastructure.wholesale.column_names import (
     DataProductColumnNames,
 )
-from tests.data_seeding import (
+from tests.settlement_reports.data_seeding import (
     standard_wholesale_fixing_scenario_data_generator,
 )
