+++ source/geh_settlement_report/tests/settlement_reports/domain/time_series_points/test_prepare_for_csv.py
@@ -7,8 +7,8 @@ from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql.functions import monotonically_increasing_id
 from pyspark.sql.types import DecimalType
 
-import tests.test_factories.default_test_data_spec as default_data
-import tests.test_factories.metering_point_time_series_factory as time_series_points_factory
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.metering_point_time_series_factory as time_series_points_factory
 from geh_settlement_report.settlement_reports.domain.time_series_points.prepare_for_csv import (
     prepare_for_csv,
 )
