+++ source/geh_settlement_report/tests/settlement_reports/domain/utils/test_system_operator_filter.py
@@ -3,10 +3,10 @@ from datetime import datetime
 import pytest
 from pyspark.sql import SparkSession
 
-import tests.test_factories.charge_link_periods_factory as charge_link_periods_factory
-import tests.test_factories.charge_price_information_periods_factory as charge_price_information_periods_factory
-import tests.test_factories.default_test_data_spec as default_data
-import tests.test_factories.metering_point_time_series_factory as time_series_points_factory
+import tests.settlement_reports.test_factories.charge_link_periods_factory as charge_link_periods_factory
+import tests.settlement_reports.test_factories.charge_price_information_periods_factory as charge_price_information_periods_factory
+import tests.settlement_reports.test_factories.default_test_data_spec as default_data
+import tests.settlement_reports.test_factories.metering_point_time_series_factory as time_series_points_factory
 from geh_settlement_report.settlement_reports.domain.time_series_points.system_operator_filter import (
     filter_time_series_points_on_charge_owner,
 )
