+++ source/geh_settlement_report/tests/settlement_reports/test_factories/default_test_data_spec.py
@@ -13,22 +13,22 @@ from geh_settlement_report.settlement_reports.infrastructure.wholesale.data_valu
 from geh_settlement_report.settlement_reports.infrastructure.wholesale.data_values.settlement_method import (
     SettlementMethodDataProductValue,
 )
-from tests.test_factories.amounts_per_charge_factory import AmountsPerChargeRow
-from tests.test_factories.charge_link_periods_factory import ChargeLinkPeriodsRow
-from tests.test_factories.charge_price_information_periods_factory import (
+from tests.settlement_reports.test_factories.amounts_per_charge_factory import AmountsPerChargeRow
+from tests.settlement_reports.test_factories.charge_link_periods_factory import ChargeLinkPeriodsRow
+from tests.settlement_reports.test_factories.charge_price_information_periods_factory import (
     ChargePriceInformationPeriodsRow,
 )
-from tests.test_factories.charge_price_points_factory import ChargePricePointsRow
-from tests.test_factories.energy_factory import EnergyTestDataSpec
-from tests.test_factories.latest_calculations_factory import LatestCalculationsPerDayRow
-from tests.test_factories.metering_point_periods_factory import MeteringPointPeriodsRow
-from tests.test_factories.metering_point_time_series_factory import (
+from tests.settlement_reports.test_factories.charge_price_points_factory import ChargePricePointsRow
+from tests.settlement_reports.test_factories.energy_factory import EnergyTestDataSpec
+from tests.settlement_reports.test_factories.latest_calculations_factory import LatestCalculationsPerDayRow
+from tests.settlement_reports.test_factories.metering_point_periods_factory import MeteringPointPeriodsRow
+from tests.settlement_reports.test_factories.metering_point_time_series_factory import (
     MeteringPointTimeSeriesTestDataSpec,
 )
-from tests.test_factories.monthly_amounts_per_charge_factory import (
+from tests.settlement_reports.test_factories.monthly_amounts_per_charge_factory import (
     MonthlyAmountsPerChargeRow,
 )
-from tests.test_factories.total_monthly_amounts_factory import TotalMonthlyAmountsRow
+from tests.settlement_reports.test_factories.total_monthly_amounts_factory import TotalMonthlyAmountsRow
 
 DEFAULT_FROM_DATE = datetime(2024, 1, 1, 23)
 DEFAULT_TO_DATE = DEFAULT_FROM_DATE + timedelta(days=1)
